Introduce AbstractRepository and use it frfom within the annotation processor

+++ integration-tests/src/test/java/arez/integration/IdentifiableIntegrationTest.java
@@ -17,7 +17,7 @@ public class IdentifiableIntegrationTest
     final Model1 model = new IdentifiableIntegrationTest_Arez_Model1();
 
     assertEquals( model instanceof Identifiable, true );
-    assertEquals( Identifiable.getArezId( model ), 0L );
+    assertEquals( Identifiable.getArezId( model ), (Long) 0L );
     assertEquals( Identifiable.asIdentifiable( model ), model );
   }
 
@@ -28,7 +28,7 @@ public class IdentifiableIntegrationTest
     final Model2 model = new IdentifiableIntegrationTest_Arez_Model2( 33 );
 
     assertEquals( model instanceof Identifiable, true );
-    assertEquals( Identifiable.getArezId( model ), 33 );
+    assertEquals( Identifiable.getArezId( model ), (Integer) 33 );
     assertEquals( Identifiable.asIdentifiable( model ), model );
   }
